refactor: :recycle: Renomeado arquivos de componentes utilizados no detalhe da planta
