+++ src/entities/detalhe-planta/condicoes-ideais-planta-section.tsx
