+++ src/entities/detalhe-planta/dificuldade-planta-section.tsx
