+++ src/entities/detalhe-planta/planta-imagem-section.tsx
